Change nginx priority to 980
This commit fixes an issue where it seems docker is affected by the change in priority to 1000 as there seems to be some related dependencies which makes apt pull in upstream docker package because it is newer then our fork atm. Changing to 980 has no functional change but just makes sure local fork of docker has more priority.